pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "x11-input-monitor"
version = "0.1.0"
description = "Demonstrationsprojekt zum Monitoring von X11-Zwischenablage und Tastatureingaben"
authors = [
    { name = "Thoms", email = "user@example.com" },
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-xlib>=0.33",
    "pyperclip>=1.8.2",
    "urwid>=2.2.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Topic :: System :: Monitoring",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/x11-input-monitor"
"Bug Tracker" = "https://github.com/yourusername/x11-input-monitor/issues"

[project.optional-dependencies]
dev = [
    "ruff"
]

[project.scripts]
x11-input-monitor = "x11_input_monitor.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["x11_input_monitor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=x11_input_monitor --cov-report=term-missing"

[tool.coverage.run]
source = ["x11_input_monitor"]
omit = ["**/__main__.py"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
